requirements.txt
google-generativeai>=0.7.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
chainlit>=1.3.0
msal>=1.30.0
httpx>=0.27
python-dateutil>=2.9
pydantic>=2.7
python-dotenv>=1.0.0